actualizacion

+++ src/app/modules/tasks/tareas/delete-tarea/delete-tarea.component.ts
@@ -2,6 +2,7 @@ import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { ToastrService } from 'ngx-toastr';
 import { TareaService } from '../service/tarea.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-delete-tarea',
@@ -24,17 +25,53 @@ export class DeleteTareaComponent {
   ngOnInit(): void {
   }
 
-  delete(){
-    
-    this.tareasService.deleteTarea(this.TAREA_SELECTED.id).subscribe((resp:any) => {
-      console.log(resp);
-      if(resp.message == 403){
-        this.toast.error("Validación",resp.message_text);
-      }else{
-        this.toast.success("Exito","La tarea se elimino correctamente");
-        this.TareaD.emit(resp.message);
-        this.modal.close();
+  delete() {
+    Swal.fire({
+      icon: 'warning',
+      title: '¿Eliminar tarea?',
+      text: 'Esta acción no se puede deshacer',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar',
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6'
+    }).then(result => {
+
+      if (result.isConfirmed && typeof this.TAREA_SELECTED.id === 'number') {
+
+        this.tareasService.deleteTarea(this.TAREA_SELECTED.id).subscribe((resp: any) => {
+          console.log(resp);
+
+          if (resp.message == 403) {
+            Swal.fire({
+              icon: 'error',
+              title: 'Validación',
+              text: resp.message_text,
+              timer: 3500,
+              showConfirmButton: false,
+              toast: true,
+              position: 'top-end'
+            });
+
+          } else {
+            Swal.fire({
+              icon: 'success',
+              title: 'Tarea eliminada',
+              text: 'La tarea se eliminó correctamente',
+              timer: 3500,
+              showConfirmButton: false,
+              toast: true,
+              position: 'top-end'
+            });
+
+            this.TareaD.emit(resp.message);
+            this.modal.close();
+          }
+        });
+
       }
-    })
+
+    });
   }
+
 }
